Better NaN handling and dataframe support (#200)
* option for either Regression or Classification model

* fine tuned selection of conflict data points

* option to provide ML target variable

* use simulation_name as output dir

* data extraction from netCDF files in a separate function

* target variable and estimator used determined in main script

* should belong to previous commit

* first step towards flexible target_vars when extracting conflict (Y) data

* added class docstrings

* constructing X and Y data as dataframes instead of arrays

* no log-scale support for now, more consistent treatment of polygons w/o feature data

* removing all polygons with 1 or more NaNs

* fully pd.dataframe support implemented

* save only selected conflicts which fall in simulation period

* finetuning print output

+++ copro/evaluation.py
@@ -88,11 +88,10 @@ def polygon_model_accuracy(
 
     # - per polygon ID, compute sum of all conflict data points and add to dataframe
     if not make_proj:
-        df_count["nr_observed_conflicts"] = df.y_test.groupby(df.ID).sum()
+        df_count["nr_observed_conflicts"] = df.y_test.groupby(df.ID).sum().astype(float)
 
     # - per polygon ID, compute sum of all conflict data points and add to dataframe
-    df_count["nr_predicted_conflicts"] = df.y_pred.groupby(df.ID).sum()
-
+    df_count["nr_predicted_conflicts"] = df.y_pred.groupby(df.ID).sum().astype(float)
     # - per polygon ID, compute average probability that conflict occurs
     df_count["min_prob_1"] = pd.to_numeric(df.y_prob_1).groupby(df.ID).min()
     df_count["probability_of_conflict"] = (
